tests/JSendWebApi.Tests/Results/JSendBadRequestResultTests.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Net;
using System.Text;
using System.Threading;
using System.Threading.Tasks;
using System.Web.Http;
using System.Web.Http.Controllers;
using FluentAssertions;
using JSendWebApi.Responses;
using JSendWebApi.Results;
using JSendWebApi.Tests.FixtureCustomizations;
using JSendWebApi.Tests.TestTypes;
using Newtonsoft.Json;
using Ploeh.AutoFixture;
using Ploeh.AutoFixture.Xunit;
using Xunit;
using Xunit.Extensions;

namespace JSendWebApi.Tests.Results
{
    public class JSendBadRequestResultTests
    {
        [Theory, JSendAutoData]
        public void IsHttpActionResult(JSendBadRequestResult<Model> result)
        {
            // Exercise system and verify outcome
            result.Should().BeAssignableTo<IHttpActionResult>();
        }

        [Theory, JSendAutoData]
        public async Task ReturnsFailJSendResponse([Frozen] Model model, JSendBadRequestResult<Model> result)
        {
            // Fixture setup
            var jsendFail = JsonConvert.SerializeObject(new FailJSendResponse<Model>(model));
            // Exercise system
            var message = await result.ExecuteAsync(new CancellationToken());
            // Verify outcome
            var content = await message.Content.ReadAsStringAsync();
            content.Should().Be(jsendFail);
        }

        [Theory, JSendAutoData]
        public async Task StatusCodeIs400(JSendBadRequestResult<Model> result)
        {
            // Exercise system
            var message = await result.ExecuteAsync(new CancellationToken());
            // Verify outcome
            message.StatusCode.Should().Be(HttpStatusCode.BadRequest);
        }

        [Fact]
        public async Task SetsCharSetHeader()
        {
            // Fixture setup
            var fixture = new Fixture().Customize(new JSendTestConventions());
            var encoding = Encoding.ASCII;
            fixture.Inject(encoding);

            var result = fixture.Create<JSendBadRequestResult<Model>>();
            // Exercise system
            var message = await result.ExecuteAsync(new CancellationToken());
            // Verify outcome
            message.Content.Headers.ContentType.CharSet = encoding.WebName;
        }

        [Theory, JSendAutoData]
        public async Task SetsContentTypeHeader(JSendBadRequestResult<Model> result)
        {
            // Exercise system
            var message = await result.ExecuteAsync(new CancellationToken());
            // Verify outcome
            message.Content.Headers.ContentType.MediaType.Should().Be("application/json");
        }
    }
}
